refactor: better general architecture for the app
