+++ src/app/loading.tsx
@@ -1,0 +1,6 @@
+// Import(s)
+import { HomeInterfaceSkeleton } from '@/features/home';
+
+export default function Loading(): JSX.Element {
+  return <HomeInterfaceSkeleton />;
+}
