+++ src/app/tools/colors-converter/loading.tsx
@@ -1,0 +1,6 @@
+// Import(s)
+import { ColorsConverterInterfaceSkeleton } from '@/features/colors-converter';
+
+export default function Loading(): JSX.Element {
+  return <ColorsConverterInterfaceSkeleton />;
+}
