+++ src/app/tools/px-rem/loading.tsx
@@ -1,0 +1,6 @@
+// Import(s)
+import { PxRemSkeletonInterface } from '@/features/px-rem';
+
+export default function Loading(): JSX.Element {
+  return <PxRemSkeletonInterface />;
+}
