+++ src/features/colors-converter/index.ts
@@ -1,8 +1,6 @@
-// Components
-export { ColorsForm } from './components/colors-form';
-
 // Hooks
-export { ColorConverterProvider, useColorsConverter } from './hooks/color-converter-hook';
+export { ColorConverterProvider } from './hooks/color-converter-hook';
 
-// Languages
-export { COLORS_CONVERTER_LANGUAGES } from './languages/colors-converter.lng';
+// Interface(s)
+export { ColorsConverterInterface } from './interface';
+export { ColorsConverterInterfaceSkeleton } from './skeleton';
