+++ src/features/px-rem/index.ts
@@ -1,9 +1,3 @@
-// Components
-export { ConversionTable } from './components/conversion-table';
-export { PxRemForm } from './components/px-rem-form';
-
-// Data
-export { PX_REM_CONVERSION } from './data/px-rem';
-
-// Languages
-export { PX_REM_LANGUAGES } from './languages/px-rem.lng';
+// Interface(s)
+export { PxRemInterface } from './interface';
+export { PxRemSkeletonInterface } from './skeleton';
