+++ src/ui/components/layout/modules/footer.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-// #region Imports
+// #region Import(s)
 
 import Link from 'next/link';
