+++ src/ui/components/layout/modules/navbar-menu.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-// #region Imports
+// #region Import(s)
 
 import Link from 'next/link';
