+++ src/ui/components/theme/theme-toggle.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-// #region Imports
+// #region Import(s)
 
 import { useTheme } from 'next-themes';
